Remove incorrect 'on-board-processing' profiler tag
As used this tag was measuring both the transfer to/from the device and the
processing on the device and thus was overestimating the time the neuflow was
taking to process the convnet by a lot.

The tag has been updated to reflect this. The <on-board-processing> profiling
is added by the compiler.